+++ src/main/java/com/spartanullnull/otil/domain/like/controller/CommentLikeController.java
@@ -1,18 +1,23 @@
 package com.spartanullnull.otil.domain.like.controller;
 
-import com.spartanullnull.otil.domain.like.service.LikeService;
+import com.spartanullnull.otil.domain.like.service.CommentLikeService;
 import com.spartanullnull.otil.security.Impl.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/posts")
 @RequiredArgsConstructor
 public class CommentLikeController {
 
-    private final LikeService likeService;
+    private final CommentLikeService commentLikeService;
 
     // 댓글 좋아요 누르기 API
     @PostMapping("/{postId}/comments/{commentId}/like")
@@ -20,7 +25,7 @@ public class CommentLikeController {
         @PathVariable Long postId,
         @PathVariable Long commentId,
         @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        likeService.createLikeForComment(userDetails.getUser().getId(), commentId);
+        commentLikeService.createLikeForComment(userDetails.getUser().getId(), commentId);
         return ResponseEntity.ok("댓글 좋아요 성공!");
     }
 
@@ -30,14 +35,14 @@ public class CommentLikeController {
         @PathVariable Long postId,
         @PathVariable Long commentId,
         @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        likeService.deleteLikeForComment(userDetails.getUser().getId(), commentId);
+        commentLikeService.deleteLikeForComment(userDetails.getUser().getId(), commentId);
         return ResponseEntity.ok("댓글 좋아요 취소!");
     }
 
     // 댓글 좋아요 개수 조회 API
     @GetMapping("/{postId}/comments/{commentId}/count")
     public ResponseEntity<String> getLikeCountForComment(@PathVariable Long commentId) {
-        Long likeCount = likeService.getLikeCountForComment(commentId);
+        Long likeCount = commentLikeService.getLikeCountForComment(commentId);
         return ResponseEntity.ok("댓글 좋아요 : " + likeCount);
     }
 }
