+++ src/main/java/com/spartanullnull/otil/domain/like/controller/PostLikeController.java
@@ -1,18 +1,23 @@
 package com.spartanullnull.otil.domain.like.controller;
 
-import com.spartanullnull.otil.domain.like.service.LikeService;
+import com.spartanullnull.otil.domain.like.service.PostLikeService;
 import com.spartanullnull.otil.security.Impl.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/posts")
 @RequiredArgsConstructor
 public class PostLikeController {
 
-    private final LikeService likeService;
+    private final PostLikeService likeService;
 
     // 게시물 좋아요 누르기 API
     @PostMapping("/{postId}/like")
